Testing the  arcpy.da.UpdateCursor function with where_clause = "SERIAL="+str(int(arcpy.GetCount_management(pointsparsedzonegraded).getOutput(0))-1)
in order to make the parsing of the fields more robust.
We need to:
 find the first row, and get the timestamp and check if timestamp is a real time or if it is corrupted;
if it is, keep on iterating to the next row until a valid timetamp is found.

Then we get to the middle rows:
We can create a buffer of rows, stored as a list of rows that keeps an arbitrary number of rows (for example 20 last rows) and if the current timestamp is not valid, then check if the last row timestamp is valid; if it is not, then iterate to the previous one, and so on.  If all of the 20 previous rows are not valid, then we must open another search cursor and start iterating the table from the current row index - 2* the buffer #